python-service/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
youtube-transcript-api>=1.1.0
sentence-transformers==3.3.1
torch>=2.0.0
pydantic==2.10.3
python-dotenv==1.0.1
